refactor: Extract filtered transaction queries to transactions/queries
Follows the pattern that we already use in ducks/recurrence/queries
and ducks/pin/queries
